+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Email notifications (Notification API) for `Account` - [ripe-id/#4](https://github.com/ripe-tech/ripe-id/issues/4)
+* Login endpoint
 
 ### Changed